+++ random.go
@@ -7,9 +7,9 @@ import (
 
 // fonction de la marche aléatoire
 
-func randomwalk(start int, temps time.Duration) []int {
+func randomwalk(start int64, temps time.Duration) []int64 {
 	position := start
-	hist := []int{position} // trace des positions
+	hist := []int64{position} // trace des positions
 
 	endTime := time.Now().Add(temps) //quand arreter
 
@@ -18,10 +18,9 @@ func randomwalk(start int, temps time.Duration) []int {
 		if len(voisins) == 0 {
 			break
 		}
+
 		position = voisins[rand.Intn(len(voisins))]
 		hist = append(hist, position)
-
 	}
 	return hist
-
 }
